db_migration/requirements.txt
pyodbc==5.0.1
pandas==2.1.4
sqlalchemy==2.0.23
chromadb==0.4.22
python-dotenv==1.0.0
tqdm==4.66.1
sentence-transformers==2.3.1
anthropic==0.18.1
